% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutOff_attrs.R
\name{cutOff_attrs}
\alias{cutOff_attrs}
\title{Selection Cutoffs for Ranked Attributes}
\usage{
cutOff_attrs(attrs, k = 0.5)
}
\arguments{
\item{attrs}{data.frame with attributes' importance.}

\item{k}{A numeric. For \code{k > 1} it indicates how many attributes to take with the highest attribute rank (chooses k best attributes). For \code{0 < k < 1} it stands for the percent
of top attributes to take (chooses best k * 100\% of attributes).}
}
\description{
Selection Cutoffs for Ranked Attributes
}
\examples{

x = information_gain(Species ~ ., iris)
cutOff_attrs(x)
cutOff_attrs(x, k = 1)

}
\author{
Damian Skrzypiec \email{damian.j.skrzypiec@gmail.com} and Zygmunt Zawadzki \email{zygmunt@zstat.pl}
}

