% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cutOff_k}
\alias{cutOff_k}
\title{Selection Cutoffs for Ranked Attributes}
\usage{
cutOff_k(x1, x2, k)
}
\arguments{
\item{x1}{Probably vector with attributes' names.}

\item{x2}{Probably vector with attributes' ranks fo cut off.}

\item{k}{A numeric. For \code{k > 1} it indicates how many attributes to take
with the highest attribute rank (chooses k best attributes). For \code{0 < k  <=1} it stands for the percent
of top attributes to take (chooses best k * 100% of attributes). For \code{k = 0} it selects top attributes
before the biggest gap in differences of scores for neighbouring attributes
(chooses a subset of attributes which are significantly better than other.)}
}
\description{
Selection Cutoffs for Ranked Attributes
}
\examples{

x = information_gain(Species ~ ., iris)
cutOff_k(rownames(x), x[[1]], 2)

}
\author{
Damian Skrzypiec, \email{damian.j.skrzypiec@gmail.com}
}

