% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\alias{equalsizeControl}
\alias{mdlControl}
\title{Discretization}
\usage{
discretize(x, y, control = list(mdlControl(), equalsizeControl()),
  keepAll = FALSE, call = NULL)

mdlControl()

equalsizeControl(k = 10)
}
\arguments{
\item{x}{Explanatory continuous variables to be discretized or a \link{formula}.}

\item{y}{Dependent variable for supervised discretization or a \link{data.frame} when \code{x} ia a \link{formula}.}

\item{control}{\code{discretizationControl} object containing the parameters for
discretization algorithm. Possible inputs are \code{mdlControl} or \code{equalsizeControl}, so far. If passed as a list, the first element is used.}

\item{keepAll}{Logical indicating if a returned \link{data.frame} should contain non
discretized columns.}

\item{call}{Keep as \code{NULL}. Inner method parameter for consistency.}

\item{k}{Number of partitions.}
}
\description{
Discretize a range of numeric attributes in the dataset into nominal
attributes. \code{Minimum Description Length} (MDL) method is set as default
control. There is also available \code{equalsizeControl} method.
}
\examples{

# vectors
discretize(x = iris[[1]], y = iris[[5]])

# list and vector
discretize(x = list(iris[[1]], iris$Sepal.Width), y = iris$Species)

# formula input
discretize(x = Species ~ ., y = iris)
discretize(Species ~ ., iris)

\dontrun{
# Same results
library(RWeka)
Rweka_disc_out <- RWeka::Discretize(Species ~ Sepal.Length, iris)[, 1]
FSelectorRcpp_disc_out <- FSelectorRcpp::discretize(Species ~ Sepal.Length,
                                                    iris)[, 1]
table(Rweka_disc_out, FSelectorRcpp_disc_out)
# But faster method
library(microbenchmark)
microbenchmark(FSelectorRcpp::discretize(Species ~ Sepal.Length, iris),
               RWeka::Discretize(Species ~ Sepal.Length, iris))

}

}
\author{
Zygmunt Zawadzki \email{zygmunt.zawadzki@gmail.com}
}
\references{
U. M. Fayyad and K. B. Irani. Multi-Interval Discretization of
  Continuous-Valued Attributes for Classi- fication Learning. In 13th
  International Joint Conference on Uncertainly in Artificial
  Intelligence(IJCAI93), pages 1022-1029, 1993.
}

