% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_gain.R
\name{information_gain}
\alias{information_gain}
\alias{information_gain.data.frame}
\alias{information_gain.default}
\alias{information_gain.dgCMatrix}
\alias{information_gain.formula}
\alias{infotmation_gain}
\title{Entropy-based filters}
\usage{
information_gain(x, y, type = c("infogain", "gainratio", "symuncert"),
  threads = 1)

\method{information_gain}{default}(x, y, type = c("infogain", "gainratio",
  "symuncert"), threads = 1)

\method{information_gain}{data.frame}(x, y, type = c("infogain", "gainratio",
  "symuncert"), threads = 1)

\method{information_gain}{formula}(x, y, type = c("infogain", "gainratio",
  "symuncert"), threads = 1)

\method{information_gain}{dgCMatrix}(x, y, type = c("infogain", "gainratio",
  "symuncert"), threads = 1)
}
\arguments{
\item{x}{data.frame or sparse matrix with independent variables, or formula.}

\item{y}{vector with dependent variable, or data.frame if \code{x} is formula}

\item{type}{method name.}

\item{threads}{number of threads for parallel backend}
}
\value{
a data.frame containing the worth of attributes in the first column and their names as row names
}
\description{
Entropy-based filters
}
\details{
The algorithms find weights of discrete attributes basing on their correlation with continous class attribute.



 \code{type = "infogain"} is \deqn{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Class) + H(Attribute) - H(Class, Attribute)}.

 \code{type = "gainratio"} is \deqn{\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute)}}{(H(Class) + H(Attribute) - H(Class, Attribute)) / H(Attribute)}

 \code{type = "symuncert"} is \deqn{2\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute) + H(Class)}}{2 * (H(Class) + H(Attribute) - H(Class, Attribute)) / (H(Attribute) + H(Class))}
}
\examples{


irisX = iris[-5]
y = iris$Species

## data.frame interface
information_gain(irisX, y)

# formula interface
information_gain(Species ~ ., y = iris)
information_gain(Species ~ ., y = iris, type = "gainratio")
information_gain(Species ~ ., y = iris, type = "symuncert")

# sparse matrix interface
library(Matrix)
i <- c(1,3:8); j <- c(2,9,6:10); x <- 7 * (1:7)
x <- sparseMatrix(i, j, x = x)
y = c(1,1,1,1,2,2,2,2)

information_gain(x,y)
information_gain(x,y, type = "gainratio")
information_gain(x,y, type = "symuncert")

}
\author{
Zygmunt Zawadzki , \email{zygmunt.zawadzki@gmail.com}
}

