% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_formula}
\alias{to_formula}
\title{Create formula object}
\usage{
to_formula(attr, class)
}
\arguments{
\item{attr}{character vector with names of independent variables}

\item{class}{single string with dependent variable name}
}
\description{
Utility function to create formula object.
}
\examples{

to_formula(names(iris)[-5], names(iris)[5])

}

