% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutOff_attrs.R
\name{cutOff_attrs}
\alias{cutOff_attrs}
\title{Selection Cutoffs for Ranked Attributes}
\usage{
cutOff_attrs(attrs, k = 0.5, rank = NULL)
}
\arguments{
\item{attrs}{data.frame with attributes' importance or vector with attributes' names.}

\item{k}{A numeric. For \code{k > 1} it indicates how many attributes to take with the highest attribute rank (chooses k best attributes). For \code{0 < k =< 1} it stands for the percent}

\item{rank}{vector with attributes' ranks to cut off (used when attrs is a vector).

of top attributes to take (chooses best k * 100% of attributes). For \code{k = 0} it selects top attributes
before the biggest gap in differences of scores for neighbouring attributes
(chooses a subset of attributes which are significantly better than other.)}
}
\description{
Selection Cutoffs for Ranked Attributes
}
\examples{

x = information_gain(Species ~ ., iris)
cutOff_attrs(rownames(x), 2, x[[1]])
cutOff_attrs(x)

}
\author{
Damian Skrzypiec \email{damian.j.skrzypiec@gmail.com} and Zygmunt Zawadzki \email{zygmunt@zstat.pl}
}

