% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretization}
\usage{
discretize(x, y, control = mdlControl(), keepAll = FALSE)
}
\arguments{
\item{x}{The explanatory continuous variables to be discretized or formula.}

\item{y}{The dependent variable for supervised discretization}

\item{control}{The \code{control} object containing the parameters for discretisation algorithm.}

\item{keepAll}{tmp}
}
\description{
Discretize a range of numeric attributes in the dataset into nominal attributes. Discretization is by \code{Minimum Description Length} (MD)L method.
}
\examples{

discretize(iris[[1]], iris[[5]])

discretize(Species ~ ., iris)

\dontrun{
# the same results
library(RWeka)
RWeka::Discretize(Species~Sepal.Length, data = iris)[, 1] -> Rweka_disc_out
FSelectorRcpp::discretize(iris$Sepal.Length, iris$Species) ->FSelectorRcpp_disc_out
table(Rweka_disc_out,FSelectorRcpp_disc_out)
# but faster method
library(microbenchmark)
microbenchmark(discretize(iris$Sepal.Length, iris$Species),
               Discretize(Species~Sepal.Length, data = iris))

}

}
\author{
Zygmunt Zawadzki , \email{zygmunt.zawadzki@gmail.com}
}
\references{
U. M. Fayyad and K. B. Irani. Multi-Interval Discretization of Continuous-Valued Attributes for Classi-
fication Learning. In 13th International Joint Conference on Uncertainly in Artificial Intelligence(IJCAI93),
pages 1022–1029, 1993.
}

