% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize.R
\name{discretize}
\alias{discretize}
\title{Discretization}
\usage{
discretize(x, y)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\description{
Discretize a range of numeric attributes in the dataset into nominal attributes. Discretization is by MDL method.
}
\examples{

discretize(iris[[1]], iris[[5]])

discretize(Species ~ ., iris)

}

