% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_search.R
\name{exhaustive_search}
\alias{exhaustive_search}
\title{Exhaustive search}
\usage{
exhaustive_search(attributes, fun, data, subsetsSizes = length(attributes),
  singleAttr = FALSE, keepAll = TRUE, allowParallel = TRUE,
  randomSubsetsNumber = 0, ...)
}
\arguments{
\item{attributes}{character vector with attributes names}

\item{fun}{function to evaluate. See Details.}

\item{data}{data set for fun function.}

\item{subsetsSizes}{sizes of attributes subsets.}

\item{singleAttr}{singleAttr}

\item{keepAll}{keepAll}

\item{allowParallel}{allowParallel,}

\item{randomSubsetsNumber}{number of random subsets. Used only if greater than zero.}

\item{\dots}{other arguments passed to foreach}
}
\description{
The algorithm for searching atrribute subset space.
}
\examples{

# evaluator from FSelector package
evaluator = function(subset, data)
{
  library(rpart)
  k <- 5
  splits <- runif(nrow(data))
  results = sapply(1:k, function(i) {
  test.idx <- (splits >= (i - 1) / k) & (splits < i / k)
  train.idx <- !test.idx
  test <- data[test.idx, , drop=FALSE]
  train <- data[train.idx, , drop=FALSE]
  tree <- rpart(to_formula(subset, "Species"), train)
  error.rate = sum(test$Species != predict(tree, test, type="c")) / nrow(test)
  return(1 - error.rate)
   })
  return(mean(results))
}

 library(doParallel)
 registerDoParallel(cores = 2)
 system.time(exhaustive_search(names(iris)[-5], evaluator, iris))
 system.time(exhaustive_search(names(iris)[-5],
                               evaluator,
                               iris,
                               allowParallel = FALSE))
 system.time(exhaustive_search(names(iris)[-5],
                               evaluator,
                               iris,
                               allowParallel = FALSE,
                               randomSubsetsNumber = 5))

}

