% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_search.R
\name{greedy_search}
\alias{greedy_search}
\title{Greedy Search}
\usage{
greedy_search(attributes, fun, data, type = c("forward", "backward"),
  allowParallel = TRUE, ...)
}
\description{
Greedy Search
}
\examples{

library(doParallel)
registerDoParallel(cores = 2)
system.time(greedy_search(names(iris)[-5], evaluator, iris))
system.time(greedy_search(names(iris)[-5], evaluator, iris, allowParallel = FALSE))

}

