% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_gain.R
\name{information_gain}
\alias{information_gain}
\alias{infotmation_gain}
\alias{sp_information_gain}
\title{Entropy-based filters}
\usage{
information_gain(formula, data, type = c("infogain", "gainratio",
  "symuncert"), threads = 1)

sp_information_gain(x, y, type = c("infogain", "gainratio", "symuncert"))
}
\arguments{
\item{formula}{description of a model}

\item{data}{data to process}

\item{type}{method name.}

\item{threads}{no idea what is it for Zygmunt?}

\item{x}{sparse matrix}

\item{y}{dependent variable}
}
\value{
a data.frame containing the worth of attributes in the first column and their names as row names
}
\description{
Entropy-based filters
}
\details{
The algorithms find weights of discrete attributes basing on their correlation with continous class attribute.



 \code{type = "infogain"} is \deqn{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Class) + H(Attribute) - H(Class, Attribute)}.

 \code{type = "gainratio"} is \deqn{\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute)}}{(H(Class) + H(Attribute) - H(Class, Attribute)) / H(Attribute)}

 \code{type = "symuncert"} is \deqn{2\frac{H(Class) + H(Attribute) - H(Class, Attribute)}{H(Attribute) + H(Class)}}{2 * (H(Class) + H(Attribute) - H(Class, Attribute)) / (H(Attribute) + H(Class))}
}
\examples{

information_gain(Species ~ ., data = iris)
information_gain(Species ~ ., data = iris, type = "gainratio")
information_gain(Species ~ ., data = iris, type = "symuncert")

# sparse examples
library(Matrix)
i <- c(1,3:8); j <- c(2,9,6:10); x <- 7 * (1:7)
x <- sparseMatrix(i, j, x = x)
y = c(1,1,1,1,2,2,2,2)

sp_information_gain(x,y)
sp_information_gain(x,y, type = "gainratio")
sp_information_gain(x,y, type = "symuncert")

}
\author{
Zygmunt Zawadzki , \email{zygmunt.zawadzki@gmail.com}
}

