% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relief.R
\name{relief}
\alias{relief}
\title{RReliefF filter}
\usage{
relief(formula, data, neighbours.count = 5, sample.size = 10)
}
\arguments{
\item{formula}{a symbolic description of a model}

\item{data}{data to process}

\item{neighbours.count}{number of neighbours to find for every sampled instance}

\item{sample.size}{number of instances to sample}
}
\value{
a data.frame containing the worth of attributes in the first column and their names as row names


}
\description{
The algorithm finds weights of continuous and discrete attributes basing on a distance between instances.
}
\examples{

data(iris)

weights <- relief(Species~., iris, neighbours.count = 5, sample.size = 20)
print(weights)
subset <- cutoff.k(weights, 2)
f <- as.simple.formula(subset, "Species")
print(f)

}
\references{
Igor Kononenko: Estimating Attributes: Analysis and Extensions of RELIEF. In: European Conference on Machine Learning, 171-182, 1994.

Marko Robnik-Sikonja, Igor Kononenko: An adaptation of Relief for attribute estimation in regression. In: Fourteenth International Conference on Machine Learning, 296-304, 1997.
}
